Backport recently merged karafka-rdkafka fixes (#913)
* Backport recently merged karafka-rdkafka fixes

Brings rdkafka-ruby in sync with the fixes merged into karafka-rdkafka since the
last sync, preserving rdkafka-ruby's original error-handling style (no
RdkafkaError.build/build_from_c/validate!).

- Consumer GC finalizer now closes the consumer and destroys its consumer queue
- Free the librdkafka string in Consumer#cluster_id/#member_id + clusterid arity
- rd_kafka_query_watermark_offsets bound with blocking: true (releases the GVL)
- Destroy rd_kafka_conf_t on failed client creation (native_config/native_kafka)
- Metadata retry no longer leaks the native struct/topic per retried attempt
- Producer#produce removes the delivery handle on any post-register failure
- Delivery callback guarded with rescue + ensure-unlock
- Admin describe_configs/incremental_alter_configs/list_offsets parse args
  before allocating native resources (no leak on rejected input)
- to_native_tpl destroys the native list when population fails partway
- incremental_alter_configs raises on a rejected entry and frees the error object
- Admin result-count out-parameter allocated as :size_t instead of :int32

* Port regression specs for the backported fixes

Brings over the karafka-rdkafka regression coverage for each backported fix,
adapted to rdkafka-ruby (same spec helpers; original error style preserved).

Integration specs (spec/integrations, run one file at a time):
- consumer_memberid_clusterid_leak_spec, watermark_offsets_gvl_spec,
  config_native_config_leak_spec, to_native_tpl_leak_spec,
  admin_setup_leak_spec, metadata_retry_leak_spec,
  incremental_alter_configs_invalid_entry_spec,
  producer_delivery_callback_raise_spec,
  producer_delivery_handle_registry_leak_spec

Unit specs:
- callbacks_spec: admin result-count pointer is sized for a native size_t
- consumer_spec: Consumer.finalizer closes + destroys the consumer queue, and an
  end-to-end reproduction that the registered finalizer destroys the exact
  consumer-queue pointer poll_batch acquired

* CI: add --network host to debian complementary integration-spec docker steps

The 'Run integration specs (compiled flow)' and precompiled-flow steps ran the
integration specs in a nested 'docker run' without --network host, so the
backported broker-dependent integration specs could not reach the Kafka broker
on the host (Connection refused). The rspec suite steps in the same workflow
already use --network host; this matches them (and karafka-rdkafka's steps).

* CI: add --network host to alpine complementary integration-spec docker steps

Same fix as the debian workflow: the alpine x86_64 and aarch64 complementary
'Run integration specs' docker steps (compiled and precompiled flows) lacked
--network host, so the backported broker-dependent integration specs could not
reach the host Kafka broker. Matches karafka-rdkafka. The aarch64 in-container
librdkafka build step intentionally keeps no --network host (no broker needed).

* CI: exclude spec/integrations from RSpec/Output cop

The backported integration specs are standalone scripts (run via bundle exec
ruby, exit-code based) that intentionally use puts/warn to report progress, so
the RSpec/Output cop flagged them. Exclude spec/integrations from it, matching
karafka-rdkafka.

+++ lib/rdkafka/callbacks/delete_topic_handler.rb
@@ -13,7 +13,7 @@ module Rdkafka
           delete_topics_result = Rdkafka::Bindings.rd_kafka_event_DeleteTopics_result(event_ptr)
 
           # Get the number of topic results
-          pointer_to_size_t = FFI::MemoryPointer.new(:int32)
+          pointer_to_size_t = FFI::MemoryPointer.new(:size_t)
           delete_topic_result_array = Rdkafka::Bindings.rd_kafka_DeleteTopics_result_topics(delete_topics_result, pointer_to_size_t)
           delete_topic_results = TopicResult.create_topic_results_from_array(pointer_to_size_t.read_int, delete_topic_result_array)
           delete_topic_handle_ptr = Rdkafka::Bindings.rd_kafka_event_opaque(event_ptr)
